Undo move to pep440_rs 0.4.0.
pep508_rs 0.2.4 (which transitively depended on pep440_rs 0.4.0) was yanked,
but poc-monotrail wasn't updated. So now things fail to build in a different way.

+++ rye/src/sources.rs
@@ -118,7 +118,11 @@ impl From<PythonVersion> for Version {
     fn from(value: PythonVersion) -> Self {
         Version {
             epoch: 0,
-            release: vec![value.major as u64, value.minor as u64, value.patch as u64],
+            release: vec![
+                value.major as usize,
+                value.minor as usize,
+                value.patch as usize,
+            ],
             pre: None,
             post: None,
             dev: None,
@@ -132,9 +136,9 @@ impl From<PythonVersionRequest> for Version {
         Version {
             epoch: 0,
             release: vec![
-                value.major as u64,
-                value.minor.unwrap_or_default() as u64,
-                value.patch.unwrap_or_default() as u64,
+                value.major as usize,
+                value.minor.unwrap_or_default() as usize,
+                value.patch.unwrap_or_default() as usize,
             ],
             pre: None,
             post: None,
